backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
aiofiles==23.2.1
python-dotenv==1.0.1
pydantic==2.7.1
numpy==1.26.4
